// src/app/model/datosTarjetaDto.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface DatosTarjetaDto { 
    canton?: string;
    celular?: string;
    direccion?: string;
    id?: number;
    idRecidencia?: number;
    nacionalidad?: string;
    nombres?: string;
    pais?: string;
    parroquia?: string;
    provincia?: string;
}